refactor: [K-6] remove SDK sidecar runtime path (#329)
