+++ Dochi/Services/Protocols/RuntimeBridgeProtocol.swift
@@ -4,6 +4,7 @@ import Foundation
 ///
 /// `@MainActor` 근거: `runtimeState`가 SwiftUI View에서 관찰되어 런타임 상태 표시
 /// (연결 중/준비됨/오류 등)에 사용되므로 UI 스레드 격리가 필요하다.
+@available(*, deprecated, message: "Legacy SDK sidecar bridge. Use NativeAgentLoopService for runtime execution.")
 @MainActor
 protocol RuntimeBridgeProtocol {
     /// Current state of the runtime process.
